implement router and reorganize repo
